code refactoring

+++ task_manager/users/views.py
@@ -1,14 +1,23 @@
 from django.contrib import messages
 from django.contrib.auth.mixins import LoginRequiredMixin
-from django.shortcuts import redirect, render
+from django.shortcuts import get_object_or_404, redirect, render
 from django.utils.translation import gettext_lazy as _
 from django.views import View
+from django.http import HttpResponseRedirect
 
 from .forms import CreateUserForm
 from .models import Users
 
 
-# Create your views here.
+class BaseUserView(LoginRequiredMixin, View):
+    login_url = '/login/'
+
+    def dispatch(self, request, *args, **kwargs):
+        if not request.user.is_authenticated:
+            messages.error(request, _('You are not logged in! Please sign in.'))
+        return super().dispatch(request, *args, **kwargs)
+
+
 class IndexUserView(View):
     def get(self, request):
         users = Users.objects.all().order_by('id')
@@ -23,14 +32,7 @@ class IndexUserView(View):
 
 class CreateUserView(View):
     def get(self, request):
-        form = CreateUserForm()
-        return render(
-            request,
-            'user/create.html',
-            context={
-                'form': form
-            },
-        )
+        return self._render_form(request, CreateUserForm())
 
     def post(self, request):
         form = CreateUserForm(request.POST)
@@ -40,6 +42,9 @@ class CreateUserView(View):
             user.save()
             messages.success(request, _('User registered successfully'))
             return redirect('login')
+        return self._render_form(request, form)
+    
+    def _render_form(self, request, form):
         return render(
             request,
             'user/create.html',
@@ -49,43 +54,40 @@ class CreateUserView(View):
         )
     
 
-class UpdateUserView(LoginRequiredMixin, View):
-    login_url = '/login/'
-
-    def dispatch(self, request, *args, **kwargs):
-        if not request.user.is_authenticated:
-            messages.error(request, _('You are not logged in! Please sign in.'))
-        return super().dispatch(request, *args, **kwargs)
-    
+class UpdateUserView(BaseUserView):
     def get(self, request, pk):
-        auth_user_id = request.user.id
-        if auth_user_id != int(pk):
-            if not request.user.is_superuser:
-                messages.error(
-                    request, 
-                    _('You do not have permission to change another user.')
-                )
-                return redirect('users')
-        user = Users.objects.get(id=pk)
+        user = self._get_user(pk)
+        if isinstance(user, HttpResponseRedirect):
+            return user
         form = CreateUserForm(instance=user)
-        return render(
-            request,
-            'user/update.html',
-            context={
-                'form': form,
-                'user': user
-            }
-        )
+        return self._render_form(request, form, user)
     
     def post(self, request, pk):
-        user = Users.objects.get(id=pk)
+        user = self._get_user(pk)
+        if isinstance(user, HttpResponseRedirect):
+            return user
         form = CreateUserForm(request.POST, instance=user)
         if form.is_valid():
-            user = form.save(commit=False)
-            user.set_password(form.cleaned_data['password'])
-            user.save()
+            updated_user = form.save(commit=False)
+            updated_user.set_password(form.cleaned_data['password'])
+            updated_user.save()
             messages.success(request, _("User successfully changed."))
             return redirect('users')
+        return self._render_form(request, form, user)
+    
+    def _get_user(self, user_id):
+        user = get_object_or_404(Users, id=user_id)
+        auth_user_id = self.request.user.id
+
+        if auth_user_id != int(user_id) and not self.request.user.is_superuser:
+            messages.error(
+                self.request, 
+                _('You do not have permission to change another user.')
+            )
+            return redirect('users')
+        return user
+    
+    def _render_form(self, request, form, user):
         return render(
             request,
             'user/update.html',
@@ -96,24 +98,11 @@ class UpdateUserView(LoginRequiredMixin, View):
         )
     
 
-class DeleteUserView(LoginRequiredMixin, View):
-    login_url = '/login/'
-
-    def dispatch(self, request, *args, **kwargs):
-        if not request.user.is_authenticated:
-            messages.error(request, _('You are not logged in! Please sign in.'))
-        return super().dispatch(request, *args, **kwargs)
-    
+class DeleteUserView(BaseUserView):    
     def get(self, request, pk):
-        auth_user_id = request.user.id
-        if auth_user_id != int(pk):
-            if not request.user.is_superuser:
-                messages.error(
-                    request, 
-                    _('You do not have permission to modify another user.')
-                )
-                return redirect('users')
-        user = Users.objects.get(id=pk)
+        user = self._get_user(pk)
+        if isinstance(user, HttpResponseRedirect):
+            return user
         return render(
             request,
             'user/delete.html',
@@ -123,15 +112,21 @@ class DeleteUserView(LoginRequiredMixin, View):
         )
     
     def post(self, request, pk):
-        auth_user_id = request.user.id
-        if auth_user_id != int(pk):
-            if not request.user.is_superuser:
-                messages.error(
-                    request, 
-                    _('You do not have permission to modify another user.')
-                )
-                return redirect('users')
-        user = Users.objects.get(id=pk)
+        user = self._get_user(pk)
+        if isinstance(user, HttpResponseRedirect):
+            return user
         user.delete()
         messages.success(request, _('User successfully deleted'))
         return redirect('users')
+    
+    def _get_user(self, user_id):
+        user = get_object_or_404(Users, id=user_id)
+        auth_user_id = self.request.user.id
+
+        if auth_user_id != int(user_id) and not self.request.user.is_superuser:
+            messages.error(
+                self.request, 
+                _('You do not have permission to change another user.')
+            )
+            return redirect('users')
+        return user
